requirements.txt
PyQt5
PyQtWebEngine
folium
exifread
branca
ttkbootstrap==1.10.1
   Pillow>=10.0.0

